requirements.txt
aiosqlite==0.17.0
APScheduler==3.10.4
asyncmy==0.2.8
asyncodbc==0.1.1
asyncpg==0.28.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
idna==3.4
iso8601==1.1.0
loguru==0.7.0
msgpack==1.0.5
multidict==6.0.4
nonebot-adapter-onebot==2.2.4
nonebot-adapter-walleq==0.2.0
nonebot-plugin-apscheduler==0.3.0
nonebot-plugin-localstore==0.5.0
nonebot-plugin-tortoise-orm==0.1.1
nonebot2==2.0.1
pydantic==1.10.12
pygtrie==2.5.0
pyodbc==4.0.39
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
tortoise-orm==0.20.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
websockets==11.0.3
win32-setctime==1.1.0
yarl==1.9.2
